requirements.txt
streamlit
numpy
pandas
ccxt
python-telegram-bot
python-dotenv
matplotlib
